+++ src/api/controllers/user.controller.ts
@@ -89,7 +89,7 @@ exports.update = (req: Request, res: Response, next: NextFunction) => {
 /**
  * Get user list
  * @public
- * @example GET http://localhost:3009/v1/users?role=admin&limit=5&offset=0&sort=email:desc,createdAt
+ * @example GET /v1/users?role=admin&limit=5&offset=0&sort=email:desc,createdAt
  */
 exports.list = async (req: Request, res: Response, next: NextFunction) => {
   try {
@@ -105,7 +105,7 @@ exports.list = async (req: Request, res: Response, next: NextFunction) => {
  * Get user's notes.
  * NOTE: Any logged in user can get a list of notes of any user.
  * @public
- * @example GET http://localhost:3009/v1/users/USERID/notes
+ * @example GET /v1/users/USERID/notes
  */
 exports.listUserNotes = async (req: Request, res: Response, next: NextFunction) => {
   try {
@@ -121,7 +121,7 @@ exports.listUserNotes = async (req: Request, res: Response, next: NextFunction) 
 
 /**
  * Add a note.
- * @example POST http://localhost:3009/v1/users/USERID/notes - payload { title, note }
+ * @example POST /v1/users/USERID/notes - payload { title, note }
  */
 exports.createNote = async (req: Request, res: Response, next: NextFunction) => {
   const { userId } = req.params;
@@ -134,6 +134,27 @@ exports.createNote = async (req: Request, res: Response, next: NextFunction) => 
     });
     const data = await newNote.save();
     apiJson({ req, res, data, model: UserNote });
+  } catch (e) {
+    next(e);
+  }
+};
+
+/**
+ * Read a user's note.
+ * NOTE: Any logged in user can get a list of notes of any user.
+ * @public
+ * @example GET /v1/users/USERID/notes/NOTEID
+ */
+exports.readNote = async (req: Request, res: Response, next: NextFunction) => {
+  const { userId, noteId } = req.params;
+  const { _id } = req.route.meta.user;
+  const currentUserId = _id.toString();
+  if (userId !== currentUserId) {
+    return next(); // only logged in user can delete her own notes
+  }
+  try {
+    const data = await UserNote.findOne({ user: new ObjectId(userId), _id: new ObjectId(noteId) });
+    apiJson({ req, res, data });
   } catch (e) {
     next(e);
   }