+++ ui/src/App.tsx
@@ -2,6 +2,7 @@ import React from 'react';
 import { BrowserRouter as Router, Routes, Route } from 'react-router-dom';
 import { Login } from './components/Login/Login';
 import { Home } from './components/Home/Home';
+import ItemView from './components/ItemView/ItemView';
 
 function App() {
   return (
@@ -9,6 +10,7 @@ function App() {
       <Routes>
         <Route path="/" element={<Login />} />
         <Route path="/home" element={<Home />} />
+        <Route path="/item/:itemId" element={<ItemView />} />
       </Routes>
     </Router>
   );