+++ ui/src/components/Home/Home.tsx
@@ -1,4 +1,5 @@
 import React from 'react';
+import { Link } from 'react-router-dom';
 import ItemView from '../ItemView/ItemView';
 import { apiGet, apiPost, apiDelete, getLoginData } from '../../utils/apiUtil';
 
@@ -35,10 +36,10 @@ export function Home() {
     await fetchData();
   };
 
-  if (selectedItem) {
-    // TODO: use react router to redirect to /item/itemId URL.
-    return <ItemView item={selectedItem} onBackClick={() => setSelectedItem(null)} />;
-  }
+  // if (selectedItem) {
+  //   // TODO: use react router to redirect to /item/itemId URL.
+  //   return <ItemView item={selectedItem} onBackClick={() => setSelectedItem(null)} />;
+  // }
   return (
     <div className="p-5">
       <h3>Current user: {loginData?.data?.user?.email}</h3>
@@ -52,8 +53,8 @@ export function Home() {
         </div>
 
         {items.map((item: any) => (
-          <div className="border rounded p-4 text-sm mt-4 mr-4" onClick={() => setSelectedItem(item)}>
-            {item.note}{' '}
+          <div className="border rounded p-4 text-sm mt-4 mr-4">
+            <Link to={`/item/${item.id}`}>{item.note} </Link>
             <a href="javascript:;" onClick={() => onClickDelete(item)}>
               ✕
             </a>